Extract common layout component
