+++ storybook/src/pages/amsterdam/ArticlePage/ArticlePage.stories.tsx
@@ -3,11 +3,8 @@
  * Copyright Gemeente Amsterdam
  */
 
-import type { FooterProps } from '@amsterdam/design-system-react'
 import { Meta, StoryObj } from '@storybook/react'
-import type { ReactElement } from 'react'
 import { ArticlePage } from './ArticlePage'
-import * as FooterStories from '../../../components/Footer/Footer.stories'
 import * as ImageStories from '../../../components/Image/Image.stories'
 import { commonMeta } from '../common/config'
 
@@ -16,7 +13,6 @@ const meta = {
   title: 'Pages/Amsterdam.nl/Article Page',
   component: ArticlePage,
   args: {
-    footer: FooterStories.Default.args?.children as ReactElement<FooterProps>,
     heading: 'Ontvang uw paspoort of ID-kaart in de zomer gratis thuis',
     imageSrc: ImageStories.LazyLoading.args?.src,
     lead: 'Niet meer naar het Stadsloket, maar thuis of op het werk uw nieuwe paspoort ontvangen. In juni en juli kan dat. Deze zomer bezorgt de gemeente reisdocumenten gratis aan huis. We doen dat om de drukte in de Stadsloketten te verminderen. En u hoeft maar 1 keer naar het Stadsloket.',
