+++ storybook/src/pages/amsterdam/ArticlePage/ArticlePage.tsx
@@ -1,13 +1,8 @@
-import { Footer, Screen, SkipLink } from '@amsterdam/design-system-react'
-import type { FooterProps } from '@amsterdam/design-system-react'
-import { ReactElement } from 'react'
 import { ArticleBody } from './ArticleBody'
 import { ArticleBreadcrumb } from './ArticleBreadcrumb'
 import { ArticleHeader } from './ArticleHeader'
-import { AppHeader } from '../common/AppHeader'
 
 export type ArticlePageProps = {
-  footer: ReactElement<FooterProps>
   heading: string
   imageSrc: string
   lead: string
@@ -17,7 +12,6 @@ export type ArticlePageProps = {
 }
 
 export const ArticlePage = ({
-  footer,
   heading,
   imageSrc,
   lead,
@@ -26,20 +20,15 @@ export const ArticlePage = ({
   spotlightLinkLabel,
 }: ArticlePageProps) => (
   <>
-    <SkipLink href="#main">Direct naar inhoud</SkipLink>
-    <Screen maxWidth="wide">
-      <AppHeader />
-      <ArticleBreadcrumb />
-      <main id="main">
-        <ArticleHeader heading={heading} imageSrc={imageSrc} />
-        <ArticleBody
-          lead={lead}
-          paragraph1={paragraph1}
-          spotlightHeading={spotlightHeading}
-          spotlightLinkLabel={spotlightLinkLabel}
-        />
-      </main>
-      <Footer>{footer}</Footer>
-    </Screen>
+    <ArticleBreadcrumb />
+    <main id="main">
+      <ArticleHeader heading={heading} imageSrc={imageSrc} />
+      <ArticleBody
+        lead={lead}
+        paragraph1={paragraph1}
+        spotlightHeading={spotlightHeading}
+        spotlightLinkLabel={spotlightLinkLabel}
+      />
+    </main>
   </>
 )
