+++ storybook/src/pages/amsterdam/HomePage/HomePage.tsx
@@ -1,29 +1,14 @@
-import { Footer, Overlap, Screen, SkipLink } from '@amsterdam/design-system-react'
-import type { FooterProps } from '@amsterdam/design-system-react'
-import { ReactElement } from 'react'
+import { Overlap } from '@amsterdam/design-system-react'
 import { HomeNews } from './HomeNews'
 import { HomeSpotlight } from './HomeSpotlight'
 import { HomeTopTasks } from './HomeTopTasks'
 import { Default as OverlapStory } from '../../../components/Overlap/Overlap.stories'
-import SkipLinkStoryMeta from '../../../components/SkipLink/SkipLink.stories'
-import { AppHeader } from '../common/AppHeader'
 
-type HomePageProps = {
-  footer: ReactElement<FooterProps>
-}
-
-export const HomePage = ({ footer }: HomePageProps) => (
+export const HomePage = () => (
   <>
-    <SkipLink href="#main">{SkipLinkStoryMeta.args.children}</SkipLink>
-    <Screen maxWidth="wide">
-      <AppHeader />
-      <main id="main">
-        <Overlap>{OverlapStory.args?.children}</Overlap>
-        <HomeTopTasks />
-        <HomeSpotlight />
-        <HomeNews />
-      </main>
-      <Footer>{footer}</Footer>
-    </Screen>
+    <Overlap>{OverlapStory.args?.children}</Overlap>
+    <HomeTopTasks />
+    <HomeSpotlight />
+    <HomeNews />
   </>
 )
