--- storybook/src/pages/amsterdam/common/config.ts
@@ -1,11 +1,0 @@
-export const commonMeta = {
-  argTypes: {
-    footer: {
-      table: { disable: true },
-    },
-  },
-  parameters: {
-    layout: 'fullscreen',
-    themes: { themeOverride: 'Spacious' },
-  },
-}
